// src/api/gen/collateral-v0.ts
/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * Collateral API (v0)
 * API for collateral estimation (v0)
 * OpenAPI spec version: v0.1.0
 */
import { axiosClientCollateralV0 } from '../client/axios/collateral-v0';

export type EstimateCollateralCollateralV0Params = {
  from_chain_id: string;
  to_chain_id: string;
  from_token_address: string;
  to_token_address: string;
};

export type ValidationErrorCollateralV0LocItem = string | number;

export interface ValidationErrorCollateralV0 {
  loc: ValidationErrorCollateralV0LocItem[];
  msg: string;
  type: string;
}

export interface HTTPValidationErrorCollateralV0 {
  detail?: ValidationErrorCollateralV0[];
}

export interface CollateralEstimateCollateralV0 {
  dashboard_url: string;
  from_price_change: string;
  increase_coefficient: string;
  to_price_change: string;
  total_increase: string;
}

type SecondParameter<T extends (...args: any) => any> = Parameters<T>[1];

/**
 * Returns collateral estimate info
 * @summary Estimate collateral
 */
export const estimateCollateralCollateralV0 = (
  params: EstimateCollateralCollateralV0Params,
  options?: SecondParameter<typeof axiosClientCollateralV0>,
) => {
  return axiosClientCollateralV0<CollateralEstimateCollateralV0>(
    { url: `/api/v0/collateral/estimate`, method: 'GET', params },
    options,
  );
};

export type EstimateCollateralCollateralV0Result = NonNullable<
  Awaited<ReturnType<typeof estimateCollateralCollateralV0>>
>;